Made sure volume value type was changed to long since it was throwing an exception for certain symbols

+++ src/Models/ChartData.cs
@@ -54,5 +54,5 @@ public class ChartInfo
 
     public List<double> CloseList { get; set; } = [];
 
-    public List<double> VolumeList { get; set; } = [];
+    public List<long> VolumeList { get; set; } = [];
 }
